requirements.txt
Flask>=3.0.0
PyYAML>=6.0.0
gunicorn>=22.0.0
python-docx>=1.1.0
pypdf>=4.2.0
